move to cmd bar
